requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
openai
smolagents
litellm
matplotlib
pandas
numpy
statsmodels
xlrd
openpyxl